cjxl: take exif orientation into account
This makes sure that the exif orientation is taken into account except
for the jpeg reconstruction case (which is planned next).
