+++ lib/extras/exif.h
@@ -10,10 +10,14 @@
 
 #include <vector>
 
+#include "jxl/codestream_header.h"
+
 namespace jxl {
 
 // Sets the Exif orientation to the identity, to avoid repeated orientation
 void ResetExifOrientation(std::vector<uint8_t>& exif);
+
+void InterpretExif(const std::vector<uint8_t>& exif, JxlBasicInfo* info);
 
 }  // namespace jxl
